move trailing stop order helper to a separate file
